Proposal: Tighten validation rules
Previously we've considered two validation rules: "require the certificate to validate for the SNI" or "don't require the certificate to validate for the SNI".  The former prevents us from setting the SNI to an unrelated domain, but the latter allows an active adversary to observe the CERTIFICATE_REQUEST frame and learn the true destination.

This change splits the difference, betting on TLS 1.3 to hide the destination domain's certificate if the SNI is an unrelated domain.